requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.1
python-docx==0.8.11
python-multipart==0.0.6
pydantic==2.5.0
tiktoken==0.5.1
